src/nlp-turnover-training/requirements.txt
absl-py==0.7.1
aspy.yaml==1.3.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
blis==0.2.4
cachetools==3.1.1
certifi==2019.6.16
cfgv==2.0.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
cymem==2.0.2
en-core-web-sm==2.1.0
filelock==3.0.12
gast==0.2.2
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-translate==1.5.0
google-pasta==0.1.7
googleapis-common-protos==1.6.0
googletrans==2.4.0
grpcio==1.22.0
h5py==2.9.0
identify==1.4.5
idna==2.8
importlib-metadata==0.18
isort==4.3.20
joblib==0.13.2
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.4
Markdown==3.1.1
mccabe==0.6.1
mpmath==1.1.0
murmurhash==1.0.2
nltk==3.4.3
nodeenv==1.3.3
numpy==1.16.4
packaging==19.0
pandas==0.24.2
plac==0.9.6
pluggy==0.12.0
pre-commit==1.17.0
preshed==2.0.1
protobuf==3.8.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pylint==2.3.1
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
spacy==2.1.4
srsly==0.0.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textblob==0.15.3
thinc==7.0.4
toml==0.10.0
tox==3.13.2
tqdm==4.32.2
translate==3.5.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
wasabi==0.2.2
Werkzeug==0.15.4
wrapt==1.11.1
zipp==0.5.1
